@@ -1,11 +1,15 @@
 import 'package:flutter/material.dart';
 
 class BurgerOffer extends StatelessWidget {
-  final url;
-  final width;
-  final height;
+  final String url;
+  final double width;
+  final double height;
 
-  BurgerOffer({this.url, this.width, this.height});
+  const BurgerOffer(
+      {super.key,
+      required this.url,
+      required this.width,
+      required this.height});
 
   @override
   Widget build(BuildContext context) {
@@ -17,7 +21,7 @@ class BurgerOffer extends StatelessWidget {
             height: height,
             width: width / 3,
             child: ClipRRect(
-              borderRadius: BorderRadius.all(
+              borderRadius: const BorderRadius.all(
                 Radius.circular(10),
               ),
               child: Image.network(
@@ -30,27 +34,36 @@ class BurgerOffer extends StatelessWidget {
             child: Align(
               alignment: Alignment.topLeft,
               child: Padding(
-                padding: EdgeInsets.fromLTRB(12.0, 12.0, 0, 0),
+                padding: const EdgeInsets.fromLTRB(12.0, 12.0, 0, 0),
                 child: Column(
                   children: [
                     Padding(
-                      padding: EdgeInsets.only(bottom: 12.0),
+                      padding: const EdgeInsets.only(bottom: 12.0),
                       child: Text(
                         'Cheese Burger',
-                        style: Theme.of(context).textTheme.bodyText1.copyWith(
-                            fontWeight: FontWeight.bold, color: Colors.white),
+                        style: Theme.of(context)
+                            .textTheme
+                            .headlineMedium!
+                            .copyWith(
+                                fontWeight: FontWeight.bold,
+                                color: Colors.white),
                       ),
                     ),
                     ElevatedButton(
-                        style: ButtonStyle(
-                            backgroundColor:
-                                MaterialStateProperty.all<Color>(Colors.green)),
-                        child: Text(
-                          '\$4.15',
-                          style: Theme.of(context).textTheme.bodyText1.copyWith(
-                              fontWeight: FontWeight.bold, color: Colors.white),
-                        ),
-                        onPressed: () {}),
+                      style: ButtonStyle(
+                          backgroundColor:
+                              MaterialStateProperty.all<Color>(Colors.green)),
+                      child: Text(
+                        '\$4.15',
+                        style: Theme.of(context)
+                            .textTheme
+                            .headlineSmall!
+                            .copyWith(
+                                fontWeight: FontWeight.bold,
+                                color: Colors.white),
+                      ),
+                      onPressed: () {},
+                    ),
                   ],
                 ),
               ),
