@@ -72,9 +72,9 @@ class IngredientSheet extends StatelessWidget {
               children: [
                 ..._chossenIngredients
                     .map((ingredient) => IngredientWidget(
-                          name: ingredient['name'],
-                          url: ingredient['url'],
-                          quantity: ingredient['quantity'],
+                          name: ingredient['name'] as String,
+                          url: ingredient['url'] as String,
+                          quantity: ingredient['quantity'] as int,
                         ))
                     .toList(),
               ],
